Split storage.rs into submodules
